chore(auth): move login related elements into single folder under layouts
