+++ ui/src/layouts/mod.rs
@@ -1,9 +1,8 @@
 pub mod chats;
 pub mod community;
-pub mod create_account;
 pub mod friends;
 pub mod loading;
+pub mod log_in;
 pub mod settings;
 pub mod slimbar;
 pub mod storage;
-pub mod unlock;
